+++ lib/presentation/pages/repo/components/readme_markdown.dart
@@ -12,10 +12,50 @@ import 'package:lottie/lottie.dart';
 import 'package:markdown/markdown.dart';
 
 import '../../../../domain/repositories/repo/entities/repo.dart';
+import '../../../../domain/repositories/repo/repo_repository.dart';
 import '../../../../utils/assets/assets.gen.dart';
 import '../../../../utils/logger.dart';
 import '../../../components/launch_url_state.dart';
-import 'readme_content.dart';
+
+/// READMEコンテンツプロバイダー（Family）
+final readmeContentProviderFamily = StateNotifierProvider.family
+    .autoDispose<ReadmeContentNotifier, AsyncValue<String>, Repo>(
+  (ref, repo) {
+    logger.i(
+      'Create ReadmeContentNotifier: fullName = ${repo.fullName}, '
+      'defaultBranch = ${repo.defaultBranch}',
+    );
+    return ReadmeContentNotifier(
+      ref.read,
+      repo: repo,
+    );
+  },
+);
+
+/// READMEコンテンツNotifier
+class ReadmeContentNotifier extends StateNotifier<AsyncValue<String>> {
+  ReadmeContentNotifier(
+    Reader read, {
+    required this.repo,
+  })  : _repoRepository = read(repoRepositoryProvider),
+        super(const AsyncValue.loading()) {
+    _get();
+  }
+
+  final RepoRepository _repoRepository;
+
+  /// リポジトリEntity
+  final Repo repo;
+
+  Future<void> _get() async {
+    state = const AsyncValue.loading();
+    state = await AsyncValue.guard(() async {
+      return _repoRepository.getReadme(
+        repo: repo,
+      );
+    });
+  }
+}
 
 /// READMEのMarkdown表示
 class ReadmeMarkdown extends ConsumerWidget {
