+++ lib/presentation/pages/repo/components/search_repos_sort_selector_bottom_sheet.dart
@@ -5,11 +5,31 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../../../../domain/repositories/app_data/app_data_repository.dart';
 import '../../../../domain/repositories/repo/entities/search_repos_sort.dart';
 import '../../../../localizations/strings.g.dart';
 import '../../../../utils/logger.dart';
 import 'search_repos_order_toggle_button.dart';
-import 'search_repos_sort.dart';
+
+/// リポジトリ検索用ソート値プロバイダー
+final searchReposSortProvider = StateProvider<SearchReposSort>(
+  (ref) {
+    final appDataRepository = ref.read(appDataRepositoryProvider);
+    return appDataRepository.getSearchReposSort();
+  },
+);
+
+/// リポジトリ検索用ソート値更新メソッドプロバイダー
+final searchReposSortUpdater = Provider(
+  (ref) {
+    final notifier = ref.read(searchReposSortProvider.notifier);
+    final appDataRepository = ref.read(appDataRepositoryProvider);
+    return (SearchReposSort sort) {
+      appDataRepository.setSearchReposSort(sort);
+      notifier.state = appDataRepository.getSearchReposSort();
+    };
+  },
+);
 
 /// リポジトリ検索用ソート選択ボトムシート
 class SearchReposSortSelectorBottomSheet extends ConsumerWidget {
