+++ lib/presentation/pages/repo/components/selected_repo.dart
@@ -3,11 +3,15 @@
 // found in the LICENSE file.
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:go_router/go_router.dart';
 
 import '../../../../domain/repositories/repo/entities/repo.dart';
 import '../../../../domain/repositories/repo/repo_repository.dart';
 import '../../../../utils/logger.dart';
-import 'selected_repo_parameter.dart';
+import '../repo_view_page.dart';
+
+part 'selected_repo.freezed.dart';
 
 /// 選択中のリポジトリプロバイダー
 final selectedRepoProvider =
@@ -26,6 +30,29 @@ final selectedRepoProviderFamily = StateNotifierProvider.family
     );
   },
 );
+
+/// 選択中のリポジトリ用パラメータ
+@freezed
+class SelectedRepoParameter with _$SelectedRepoParameter {
+  const factory SelectedRepoParameter({
+    /// オーナー名
+    required String ownerName,
+
+    /// リポジトリ名
+    required String repoName,
+
+    /// 一覧画面から渡されるリポジトリデータ
+    /// 詳細画面で再読込した場合などは null になる場合がある
+    Repo? extra,
+  }) = _SelectedRepoParameter;
+
+  factory SelectedRepoParameter.from(GoRouterState state) =>
+      SelectedRepoParameter(
+        ownerName: state.params[pageParamKeyOwnerName]!,
+        repoName: state.params[pageParamKeyRepoName]!,
+        extra: state.extra as Repo?,
+      );
+}
 
 /// 選択中のリポジトリNotifier
 class SelectedRepoNotifier extends StateNotifier<AsyncValue<Repo>> {
