+++ test/presentation/pages/repo/components/search_repos_order_toggle_button_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:github_search/domain/repositories/repo/entities/search_repos_order.dart';
 import 'package:github_search/infrastructure/github/http_client.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_order_toggle_button.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_query.dart';
@@ -17,6 +18,35 @@ void main() {
   setUp(agent.setUp);
   tearDown(agent.tearDown);
 
+  group('searchReposOrderProvider', () {
+    test('初期値は降順のはず', () async {
+      final order = agent.mockContainer().read(searchReposOrderProvider);
+      expect(order, SearchReposOrder.desc);
+    });
+  });
+  group('searchReposOrderUpdater', () {
+    test('オーダー値を変更できるはず', () async {
+      final updater = agent.mockContainer().read(searchReposOrderUpdater);
+
+      // 昇順に変更する
+      updater(SearchReposOrder.asc);
+
+      // 昇順のはず
+      expect(
+        agent.mockContainer().read(searchReposOrderProvider),
+        SearchReposOrder.asc,
+      );
+
+      // 降順に変更する
+      updater(SearchReposOrder.desc);
+
+      // 降順のはず
+      expect(
+        agent.mockContainer().read(searchReposOrderProvider),
+        SearchReposOrder.desc,
+      );
+    });
+  });
   group('SearchReposOrderToggleButton', () {
     testWidgets('オーダーボタン押下で昇順降順を切り替えられるはず', (tester) async {
       await tester.pumpWidget(
