+++ test/presentation/pages/repo/components/search_repos_sort_selector_bottom_sheet_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:github_search/domain/repositories/repo/entities/search_repos_sort.dart';
 import 'package:github_search/infrastructure/github/http_client.dart';
 import 'package:github_search/localizations/strings.g.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_order_toggle_button.dart';
@@ -34,6 +35,53 @@ void main() {
   setUp(agent.setUp);
   tearDown(agent.tearDown);
 
+  group('searchReposSortProvider', () {
+    test('初期値はベストマッチのはず', () async {
+      final sort = agent.mockContainer().read(searchReposSortProvider);
+      expect(sort, SearchReposSort.bestMatch);
+    });
+  });
+  group('searchReposSortUpdater', () {
+    test('ソート値を変更できるはず', () async {
+      final updater = agent.mockContainer().read(searchReposSortUpdater);
+
+      // スター数に変更する
+      updater(SearchReposSort.stars);
+
+      // スター数のはず
+      expect(
+        agent.mockContainer().read(searchReposSortProvider),
+        SearchReposSort.stars,
+      );
+
+      // フォーク数に変更する
+      updater(SearchReposSort.forks);
+
+      // フォーク数のはず
+      expect(
+        agent.mockContainer().read(searchReposSortProvider),
+        SearchReposSort.forks,
+      );
+
+      // ヘルプ数に変更する
+      updater(SearchReposSort.helpWantedIssues);
+
+      // ヘルプ数のはず
+      expect(
+        agent.mockContainer().read(searchReposSortProvider),
+        SearchReposSort.helpWantedIssues,
+      );
+
+      // ベストマッチに変更する
+      updater(SearchReposSort.bestMatch);
+
+      // ベストマッチのはず
+      expect(
+        agent.mockContainer().read(searchReposSortProvider),
+        SearchReposSort.bestMatch,
+      );
+    });
+  });
   group('SearchReposSortSelectorBottomSheet', () {
     testWidgets('各選択肢を選択してソートが変更されるはず', (tester) async {
       await tester.pumpWidget(
